# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPI Emanuel Corp API",
        "description": "API Documentation untuk Aplikasi KPI Emanuel Corp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/kpi-jenis": {
            "get": {
                "tags": [
                    "KPI Jenis Catalog"
                ],
                "summary": "Get daftar katalog KPI (untuk dropdown admin), bisa difilter per kategori unit bisnis",
                "operationId": "2008826f945ead79411faf590ac8ad09",
                "parameters": [
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "description": "Filter KPI yang relevan untuk kategori unit bisnis tertentu (Properti/PM/UMKM). Kalau tidak diisi, tampilkan semua KPI aktif.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar jenis KPI yang aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/KpiJenisResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "KPI Jenis Catalog"
                ],
                "summary": "Tambah jenis KPI baru ke katalog (owner only)",
                "operationId": "8c484fde26c7ba415e5d134e2a0f088c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode",
                                    "nama",
                                    "kategori",
                                    "satuan_default",
                                    "formula_type",
                                    "field_definitions"
                                ],
                                "properties": {
                                    "kode": {
                                        "type": "string",
                                        "example": "complaint_count"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Jumlah Komplain"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "enum": [
                                            "sales",
                                            "operasional",
                                            "keuangan",
                                            "sdm",
                                            "lainnya"
                                        ],
                                        "example": "operasional"
                                    },
                                    "satuan_default": {
                                        "type": "string",
                                        "enum": [
                                            "rupiah",
                                            "persen",
                                            "unit",
                                            "malam",
                                            "lainnya"
                                        ],
                                        "example": "unit"
                                    },
                                    "formula_type": {
                                        "type": "string",
                                        "enum": [
                                            "sum",
                                            "average",
                                            "count",
                                            "last_value",
                                            "minimize",
                                            "range",
                                            "binary"
                                        ],
                                        "example": "minimize"
                                    },
                                    "field_definitions": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "label": {
                                                    "type": "string",
                                                    "example": "Jumlah Komplain"
                                                },
                                                "tipe": {
                                                    "type": "string",
                                                    "example": "number"
                                                },
                                                "wajib": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "nilai_min": {
                                        "description": "Wajib diisi kalau formula_type = range",
                                        "type": "number",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "nilai_max": {
                                        "description": "Wajib diisi kalau formula_type = range, harus lebih besar dari nilai_min",
                                        "type": "number",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "total_milestone": {
                                        "description": "Wajib diisi kalau formula_type = binary",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "is_capped": {
                                        "description": "Batasi persentase maksimal 100%, default true",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Deskripsi tambahan KPI",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jenis KPI berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kpi-jenis/{id}": {
            "get": {
                "tags": [
                    "KPI Jenis Catalog"
                ],
                "summary": "Get detail 1 jenis KPI dari katalog",
                "operationId": "2ac8ade610687aef280b3f8d7c3ca442",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail jenis KPI",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KpiJenisResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "KPI Jenis Catalog"
                ],
                "summary": "Update jenis KPI di katalog (owner only)",
                "operationId": "46028fd002ae480c1b737577d8f34c10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Revenue Updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jenis KPI berhasil diupdate"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kpi-jenis/{id}/toggle": {
            "patch": {
                "tags": [
                    "KPI Jenis Catalog"
                ],
                "summary": "Aktifkan/nonaktifkan jenis KPI (owner only)",
                "operationId": "1b67a533936dd3e4a7eca92234e2dc5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status jenis KPI berhasil diubah"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "KpiJenisResponse": {
                "title": "KpiJenis Response Schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kode": {
                        "type": "string",
                        "example": "revenue"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Revenue"
                    },
                    "kategori": {
                        "type": "string",
                        "example": "keuangan"
                    },
                    "satuan_default": {
                        "type": "string",
                        "example": "rupiah"
                    },
                    "formula_type": {
                        "type": "string",
                        "enum": [
                            "sum",
                            "average",
                            "count",
                            "last_value",
                            "minimize",
                            "range",
                            "binary"
                        ],
                        "example": "sum"
                    },
                    "field_definitions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "label": {
                                    "type": "string",
                                    "example": "Nilai Transaksi"
                                },
                                "tipe": {
                                    "type": "string",
                                    "example": "number"
                                },
                                "wajib": {
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "nilai_min": {
                        "description": "Batas bawah skala — dipakai formula range",
                        "type": "number",
                        "example": 1,
                        "nullable": true
                    },
                    "nilai_max": {
                        "description": "Batas atas skala — dipakai formula range",
                        "type": "number",
                        "example": 5,
                        "nullable": true
                    },
                    "total_milestone": {
                        "description": "Total milestone yang harus dicapai — dipakai formula binary",
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "is_capped": {
                        "description": "Apakah persentase dibatasi maksimal 100%",
                        "type": "boolean",
                        "example": true
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Deskripsi katalog KPI",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-27T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-27T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}